AbstractRule as new root of Rule inheritance hierarchy
