+++ src/main/java/at/ac/tuwien/kr/alpha/common/rule/head/Head.java
@@ -1,0 +1,12 @@
+package at.ac.tuwien.kr.alpha.common.rule.head;
+
+/**
+ * Represents the head of a rule, i.e., either a choice or a disjunction of atoms, but not both. For normal rules the disjunction contains only one atom.
+ *
+ * Copyright (c) 2017-2019, the Alpha Team.
+ */
+public abstract class Head {
+
+	public abstract boolean isNormal();
+
+}
