+++ src/main/java/at/ac/tuwien/kr/alpha/common/rule/head/impl/ChoiceHead.java
@@ -1,11 +1,13 @@
-package at.ac.tuwien.kr.alpha.common;
+package at.ac.tuwien.kr.alpha.common.rule.head.impl;
 
 import static at.ac.tuwien.kr.alpha.Util.join;
 
 import java.util.List;
 
+import at.ac.tuwien.kr.alpha.common.ComparisonOperator;
 import at.ac.tuwien.kr.alpha.common.atoms.Atom;
 import at.ac.tuwien.kr.alpha.common.atoms.Literal;
+import at.ac.tuwien.kr.alpha.common.rule.head.Head;
 import at.ac.tuwien.kr.alpha.common.terms.Term;
 
 /**
