+++ src/main/java/at/ac/tuwien/kr/alpha/common/rule/head/impl/DisjunctiveHead.java
@@ -1,6 +1,7 @@
-package at.ac.tuwien.kr.alpha.common;
+package at.ac.tuwien.kr.alpha.common.rule.head.impl;
 
 import at.ac.tuwien.kr.alpha.common.atoms.Atom;
+import at.ac.tuwien.kr.alpha.common.rule.head.Head;
 
 import java.util.List;
