+++ src/test/java/at/ac/tuwien/kr/alpha/common/RuleTest.java
@@ -1,5 +1,6 @@
 package at.ac.tuwien.kr.alpha.common;
 
+import at.ac.tuwien.kr.alpha.common.rule.impl.Rule;
 import at.ac.tuwien.kr.alpha.grounder.parser.ProgramParser;
 import org.junit.Test;
